Day and night, off the tick nobody had to send
Day/night v0 (DECISIONS.md §open). ALPHA §1 asked for a 45-minute cycle
and NETCODE §3 files time of day on the G channel; both are answered
with no wire byte, because the snapshot header already carries the tick
and the client already smooths an estimate of it. The clock is
world::day_frac(tick) — pure, in the crate both sides share — so two
clients agree about the hour because they agree about the tick. Tick
zero boots mid-morning on purpose: every capture probe and the art
rubric score daylight frames.

The renderer is the only reader, inside rig.rs because that file owns
the coupled set (ART rule 5): one daylight() scalar drives the sun's
illuminance, the ambient lerp to a 60-lux night floor, and the cloud
deck's brightness, so the three cannot disagree. Fixed azimuth, peak at
the authored elevation, shadows off once the sun is under the ground.

Audio gets the one consumer that was blocked on this: birds are gated
to daylight, repaying the refusal birds.rs' header recorded. Gated by
tests/daynight.rs — the curve as arithmetic, plus the real system
driven on a bare App and asserted on the entities it wrote.

+++ crates/sim-core/tests/mob.rs
@@ -405,7 +405,7 @@ fn a_whole_pig_charges_and_bites() {
     assert!(full > 0, "combat fixture arms bodies");
     // Two bite periods plus a think: enough for at least one landed bite,
     // asserted on state rather than on which tick it happened.
-    for seq in 0..(MobContent::probe_fixture().def(MOB_PIG).attack_ticks as u16 * 2 + 30) {
+    for seq in 0..(MobContent::probe_fixture().def(MOB_PIG).attack_ticks * 2 + 30) {
         let frame = InputFrame {
             seq,
             ..InputFrame::default()
@@ -484,7 +484,7 @@ fn a_bite_can_kill_and_the_cause_is_the_mob() {
     w.mobs.m[slot].body = Body::at(11, ax + 2.0, az);
     // One bite from dead.
     w.players[0].hp = 1;
-    for seq in 0..(MobContent::probe_fixture().def(MOB_PIG).attack_ticks as u16 * 2 + 30) {
+    for seq in 0..(MobContent::probe_fixture().def(MOB_PIG).attack_ticks * 2 + 30) {
         let frame = InputFrame {
             seq,
             ..InputFrame::default()
